Update to 1.19.2
This is a minor release which includes security fixes to the `archive/tar`, `net/http/httputil`, and `regexp` packages, as well as bug fixes to the compiler, the linker, the runtime, and the go/types package.
See the Go 1.19.2 [milestone](https://github.com/golang/go/issues?q=milestone%3AGo1.19.2+label%3ACherryPickApproved) for details.

Signed-off-by: Algent Albrahimi <algent@protonmail.com>